Move Route utils to utils folder
