+++ src/utils/route/index.js
@@ -1,0 +1,4 @@
+import GuestRoute from "./GuestRoute";
+import ProtectedRoute from "./ProtectedRoute";
+
+export { GuestRoute, ProtectedRoute };
